// src/models/IPost.ts
export interface IPost {
  title: string;
  author: string;
  date: string;
  _content: string;
  source: string;
  raw: string;
  slug: string;
  published: string;
  updated: string;
  comments: string;
  layout: string;
  photos: string;
  link: string;
  _id: string;
  content: string;
  site: string;
  excerpt: string;
  more: string;
  path: string;
  permalink: string;
  full_source: string;
  asset_dir: string;
  tags: string;
  categories: string;
  isDraft: string;
  isDiscarded: string;
}